Add files via upload
The time conversion program allows you to convert the current time of India to five different countries. It uses the `Calendar` and `SimpleDateFormat` classes to handle time manipulation and formatting. The program retrieves the current time in India and then converts it to the time zones of the selected countries. The target time zones are predefined in the code for five countries: New York (USA), London (UK), Sydney (Australia), Tokyo (Japan), and Johannesburg (South Africa). The converted times are displayed, showing the source time (India) and the corresponding time in each target time zone. This program provides a convenient way to quickly see the current time in multiple countries relative to India.